+++ client/src/components/Reviews/ReviewList.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import styled from 'styled-components';
 
-import credentials from '../../../config.js'
+import credentials from '../../../../config.js'
 const url = 'http://localhost:3000'
 
 import Review from './Review.jsx'
